// reactive-ui-app/src/app/models/signup.model.ts
export interface SignUpModel {
    busName?: string;
    defLocnNbr?: string;
    userName?: string;
    authType?: string;
    authToken?: string;
    firstName?: string;
    lastName?: string;
    middleName?: string;
    addr1?: string;
    addr2?: string;
    addr3?: string;
    city?: string;
    state?: string;
    country?: string;
    zipCode?: string;
    locale?: string;
    userId?: string;
}